Shahar: simulate vision source config changed
