+++ src/main/java/frc/robot/poseestimator/PoseEstimatorConstants.java
@@ -2,6 +2,7 @@ package frc.robot.poseestimator;
 
 
 import edu.wpi.first.apriltag.AprilTagFields;
+import frc.robot.vision.sources.simulationsource.SimulatedSourceConfiguration;
 
 public class PoseEstimatorConstants {
 
@@ -11,7 +12,7 @@ public class PoseEstimatorConstants {
 
 	protected static final double POSE_BUFFER_SIZE_SECONDS = 2.0;
 
-	public static final AprilTagFields aprilTagField = AprilTagFields.k2024Crescendo;
+	public static final AprilTagFields APRIL_TAG_FIELD = AprilTagFields.k2024Crescendo;
 
 	public static final double[] DEFAULT_ODOMETRY_STANDARD_DEVIATIONS = {0.003, 0.003, 0.0002};
