+++ src/main/java/frc/robot/vision/sources/calibration.md
@@ -1,0 +1,2 @@
+# Calibrations
+ - Detection range etc. for each camera, in order to be able to simulate them.
