+++ src/main/java/frc/robot/vision/sources/simulationsource/SimulatedSourceConfiguration.java
@@ -1,0 +1,13 @@
+package frc.robot.vision.sources.simulationsource;
+
+import edu.wpi.first.math.geometry.Pose2d;
+
+import java.util.function.Supplier;
+
+public record SimulatedSourceConfiguration(
+	double angleNoiseScaling,
+	double transformNoiseScaling,
+	double spikesProbability,
+	double maximumSpikeMeters,
+	double detectionRangeMeters
+) {}
